updating car prices, rent dates, delete operations rest api

+++ web/src/main/java/sk/durovic/api/controllers/CarControllerRest.java
@@ -21,10 +21,12 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import org.thymeleaf.spring5.expression.Fields;
 import sk.durovic.api.dto.CarDto;
+import sk.durovic.helper.DateTimeHelper;
 import sk.durovic.httpError.NotAuthorized;
 import sk.durovic.httpError.NotFound;
 import sk.durovic.mappers.CarMapper;
 import sk.durovic.model.*;
+import sk.durovic.services.AvailabilityService;
 import sk.durovic.services.CarService;
 import sk.durovic.services.PricesService;
 
@@ -36,6 +38,7 @@ import java.util.*;
 import java.util.stream.Collectors;
 
 import static sk.durovic.helper.CarOwnerHelper.isOwnerOfCar;
+import static sk.durovic.helper.DateTimeHelper.getLocalDateTime;
 
 @RestController
 @Slf4j
@@ -47,6 +50,9 @@ public class CarControllerRest {
 
     @Autowired
     PricesService pricesService;
+
+    @Autowired
+    AvailabilityService availabilityService;
 
     private final ObjectMapper jsonData = new ObjectMapper();
 
@@ -82,24 +88,52 @@ public class CarControllerRest {
             }
         }
 
-        Car saved = carService.save(car);
-
-        JsonNode jsonNode = getJsonNodeFromCar(car);
-
-
-        return ResponseEntity.ok(jsonNode);
+        return ResponseEntity.ok(getJsonNodeFromCar(carService.save(car)));
     }
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
-    public void deleteCar(@PathVariable("id") Long id){
-        carService.deleteById(id);
+    public void deleteCar(@PathVariable("id") Long id,
+                          @RequestBody String jsonBody,
+                          @AuthenticationPrincipal UserDetails userDetails) throws JsonProcessingException {
+        Car car = carService.findById(id);
+        Set<Long> listPrices = pricesService.findAll().stream().map(Prices::getId)
+                .collect(Collectors.toSet());
+        Set<Long> listRentDates = availabilityService.findAll().stream().map(Availability::getId)
+                .collect(Collectors.toSet());
+
+        if(!isOwnerOfCar(userDetails, car))
+            throw new NotAuthorized();
+
+        JsonNode jn = jsonData.readTree(jsonBody);
+        boolean isEmptyBody=true;
+
+
+        if(!jn.path("prices").isMissingNode()){
+            long[] prices = jsonData.treeToValue(jn.path("prices"), long[].class);
+            Arrays.stream(prices).filter(listPrices::contains).forEach(pricesService::deleteById);
+            isEmptyBody=false;
+        }
+
+        if(!jn.path("rentDates").isMissingNode()){
+            long[] rentDates = jsonData.treeToValue(jn.path("rentDates"), long[].class);
+            Arrays.stream(rentDates).filter(listRentDates::contains).forEach(availabilityService::deleteById);
+            isEmptyBody=false;
+        }
+
+        if(isEmptyBody)
+            carService.deleteById(id);
     }
 
     @PutMapping("/{id}")
     public ResponseEntity<?> updateCar(@PathVariable("id") Long id,
-                                       @RequestBody String car) throws IllegalAccessException, JsonProcessingException, NoSuchFieldException {
+                                       @RequestBody String car,
+                                       @AuthenticationPrincipal UserDetails userDetails) throws IllegalAccessException, JsonProcessingException, NoSuchFieldException {
         Car updateCar = carService.findById(id);
+
+        if(!isOwnerOfCar(userDetails, updateCar))
+            throw new NotAuthorized();
+
         JsonNode updateFields = jsonData.readTree(car);
 
         /// prerobit na samostatne funkcie
@@ -125,62 +159,91 @@ public class CarControllerRest {
         if (changed)
             updateCar = carService.save(updateCar);
 
-        JsonNode jsonNode = getJsonNodeFromCar(updateCar);
-
-        return ResponseEntity.status(200).body(jsonNode);
+        return ResponseEntity.status(200).body(getJsonNodeFromCar(updateCar));
     }
 
     private Object getValue(Field field, Map.Entry<String, JsonNode> map, Car car) throws JsonProcessingException {
         Class<?> clazz = field.getType();
 
         if(map.getKey().equals("prices")){
-            Set<Prices> pricesSet = car.getPrices();
-            ArrayNode arrayNode = (ArrayNode) map.getValue();
-            Iterator<JsonNode> it = arrayNode.iterator();
-
-            while(it.hasNext()){
-                JsonNode jn = it.next();
-                Prices prices = new Prices();
-
-                if(jn.path("prices").path("days").isMissingNode() ||
-                    jn.path("prices").path("price").isMissingNode())
-                    continue;
-
-                if(!jn.path("prices").path("id").isMissingNode())
-                    prices = pricesSet.stream().filter(price -> price.getId().equals(jn.get("id").asLong()))
-                        .findFirst().orElse(Prices.builder(car).build());
-                else
-                    prices = pricesSet.stream().filter(price -> price.getDays().equals(jn.get("days").asInt()))
-                        .findFirst().orElse(Prices.builder(car).build());
-
-
-                Iterator<Map.Entry<String, JsonNode>> itPrice = jn.fields();
-
-                while(itPrice.hasNext()){
-                    Map.Entry<String, JsonNode> mapPrice = itPrice.next();
-
-                    switch (mapPrice.getKey()) {
-                        case "days":
-                            prices.setDays(mapPrice.getValue().asInt());
-                            break;
-                        case "price":
-                            prices.setPrice(mapPrice.getValue().asInt());
-                            break;
-                    }
-
-                }
-
-                if(prices.getDays()!=null && prices.getPrice()!=null)
-                    pricesSet.add(prices);
-            }
-
-            return pricesSet;
+            return getPrices(map, car);
         }
 
-
+        if(map.getKey().equals("rentDates"))
+            return getRentDates(map, car);
 
         return jsonData.treeToValue(map.getValue(), clazz);
 
+    }
+
+    private Set<Availability> getRentDates(Map.Entry<String, JsonNode> map, Car car){
+        Set<Availability> availabilitySet = car.getRentDates();
+        ArrayNode arrayNode = (ArrayNode) map.getValue();
+
+        Iterator<JsonNode> it = arrayNode.iterator();
+
+        while(it.hasNext()){
+            JsonNode jn = it.next();
+            Availability availability = new Availability();
+            availability.setCarRented(car);
+
+            if(jn.path("start").isMissingNode() || jn.path("end").isMissingNode())
+                continue;
+
+            if(!jn.path("id").isMissingNode()) {
+                availability = car.getRentDates().stream().filter(av -> av.getId().equals(jn.get("id").asLong()))
+                        .findFirst().orElse(new Availability());
+                availability.setCarRented(car);
+            }
+
+            String[] start = jn.get("start").asText().split(" ");
+            String[] end = jn.get("end").asText().split(" ");
+
+            if(start.length==1)
+                start = new String[]{start[0], ""};
+
+            if(end.length==1)
+                end = new String[]{end[0], ""};
+
+            availability.setStart(getLocalDateTime(start[0], start[1]));
+            availability.setEnd(getLocalDateTime(end[0], end[1]));
+
+            availabilitySet.add(availability);
+        }
+
+        return availabilitySet;
+    }
+
+    private Set<Prices> getPrices(Map.Entry<String, JsonNode> map, Car car) {
+        Set<Prices> pricesSet = car.getPrices();
+        ArrayNode arrayNode = (ArrayNode) map.getValue();
+
+        Iterator<JsonNode> it = arrayNode.iterator();
+
+        while(it.hasNext()){
+            JsonNode jn = it.next();
+            Prices prices = new Prices();
+            prices.setCar(car);
+
+            if(jn.path("days").isMissingNode() ||
+                jn.path("price").isMissingNode())
+                continue;
+
+            if(!jn.path("id").isMissingNode())
+                prices = pricesSet.stream().filter(price -> price.getId().equals(jn.get("id").asLong()))
+                    .findFirst().orElse(Prices.builder(car).build());
+            else
+                prices = pricesSet.stream().filter(price -> price.getDays().equals(jn.get("days").asInt()))
+                    .findFirst().orElse(Prices.builder(car).build());
+
+            prices.setDays(jn.get("days").asInt());
+            prices.setPrice(jn.get("price").asInt());
+
+            if(prices.getDays()!=null && prices.getPrice()!=null)
+                pricesSet.add(prices);
+        }
+
+        return pricesSet;
     }
 
     private JsonNode getJsonNodeFromCar(Car car) {
